% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPatientShapePalette}
\alias{getPatientShapePalette}
\title{Get shape palette for \code{\link[plotly]{plotly}}}
\usage{
getPatientShapePalette(n = NULL, x = NULL)
}
\arguments{
\item{n}{number of elements in the palette}

\item{x}{vector with elements used as names for the palette.
If factor, the levels are used, otherwise the unique elements of the vector.}
}
\value{
vector of integer values with shape
}
\description{
Note that 20 unique symbols are available at maximum.
}
\author{
Laure Cougnaud
}
